PotionSpoof rework
Spoofing clientside effects is unreliable/broken on pretty much any server.
Module rebranded as NoStatusEffects with only effect blocking capability.
