+++ src/main/java/meteordevelopment/meteorclient/systems/modules/Modules.java
@@ -437,10 +437,10 @@ public class Modules extends System<Modules> {
         add(new NoInteract());
         add(new NoMiningTrace());
         add(new NoRotate());
+        add(new NoStatusEffects());
         add(new OffhandCrash());
         add(new Portals());
         add(new PotionSaver());
-        add(new PotionSpoof());
         add(new Reach());
         add(new Rotation());
         add(new SpeedMine());
