+++ src/main/java/meteordevelopment/meteorclient/systems/modules/player/NoStatusEffects.java
@@ -1,0 +1,39 @@
+/*
+ * This file is part of the Meteor Client distribution (https://github.com/MeteorDevelopment/meteor-client).
+ * Copyright (c) Meteor Development.
+ */
+
+package meteordevelopment.meteorclient.systems.modules.player;
+
+import meteordevelopment.meteorclient.settings.*;
+import meteordevelopment.meteorclient.systems.modules.Categories;
+import meteordevelopment.meteorclient.systems.modules.Module;
+import net.minecraft.entity.effect.StatusEffect;
+
+import java.util.List;
+
+import static net.minecraft.entity.effect.StatusEffects.*;
+
+public class NoStatusEffects extends Module {
+    private final SettingGroup sgGeneral = settings.getDefaultGroup();
+
+    private final Setting<List<StatusEffect>> blockedEffects = sgGeneral.add(new StatusEffectListSetting.Builder()
+        .name("blocked-effects")
+        .description("Effects to block.")
+        .defaultValue(
+            LEVITATION.value(),
+            JUMP_BOOST.value(),
+            SLOW_FALLING.value(),
+            DOLPHINS_GRACE.value()
+        )
+        .build()
+    );
+
+    public NoStatusEffects() {
+        super(Categories.Player, "no-status-effects", "Blocks specified status effects.");
+    }
+
+    public boolean shouldBlock(StatusEffect effect) {
+        return isActive() && blockedEffects.get().contains(effect);
+    }
+}
